+++ qCali/src/main/java/com/group/exam/qna/dao/QnaDao.java
@@ -9,10 +9,10 @@ public interface QnaDao {
 	public void qnaInsert(QnaVo vo);
 	
 	public List<QnaVo> list(Criteria cri);
-	public QnaVo detail(int qnaSeq);
+	public QnaVo detail(Long qnaSeq);
 	public void qnaUpdate(QnaVo vo);
-	public void qnaDelete(int qnaSeq);
-	public void qnaRead(int qnaSeq); //조회수 증가
+	public void qnaDelete(Long qnaSeq);
+	public void qnaRead(Long qnaSeq); //조회수 증가
 	public void qnaReplyInsert(QnaVo vo);//답글
 	public int qnaTotalList(); //페이징 처리를 위한 total수
